requirements.txt
boto3==1.21.8
kubernetes==12.0.1
PyYAML==5.4
pytest-xdist==2.2.0
pydantic==1.9.1